SLVS-1910 Move interface near implementation
